improve test utils reuse
