+++ karate-core/src/main/java/com/intuit/karate/http/HttpRequestBuilder.java
@@ -23,6 +23,7 @@
  */
 package com.intuit.karate.http;
 
+import com.intuit.karate.Json;
 import com.intuit.karate.StringUtils;
 import com.intuit.karate.graal.JsArray;
 import com.intuit.karate.graal.JsValue;
@@ -228,8 +229,8 @@ public class HttpRequestBuilder implements ProxyObject {
         this.method = method;
         return this;
     }
-    
-    public HttpRequestBuilder paths(String ... paths) {
+
+    public HttpRequestBuilder paths(String... paths) {
         for (String path : paths) {
             path(path);
         }
@@ -277,6 +278,11 @@ public class HttpRequestBuilder implements ProxyObject {
 
     public HttpRequestBuilder body(Object body) {
         this.body = body;
+        return this;
+    }
+
+    public HttpRequestBuilder bodyJson(String json) {
+        this.body = Json.of(json).value();
         return this;
     }
 
@@ -409,6 +415,10 @@ public class HttpRequestBuilder implements ProxyObject {
         }
         multiPart.part(name).value(value).add();
         return this;
+    }
+    
+    public HttpRequestBuilder multiPartJson(String json) {
+        return multiPart(Json.of(json).value());
     }
 
     public HttpRequestBuilder multiPart(Map<String, Object> map) {
