+++ karate-core/src/test/java/com/intuit/karate/core/FeatureRuntimeTest.java
@@ -1,5 +1,6 @@
 package com.intuit.karate.core;
 
+import com.intuit.karate.TestUtils;
 import com.intuit.karate.match.Match;
 import com.intuit.karate.match.MatchResult;
 import java.io.File;
@@ -30,7 +31,7 @@ class FeatureRuntimeTest {
     }
 
     private FeatureRuntime run(String name, String configDir) {
-        fr = RuntimeUtils.runFeature("classpath:com/intuit/karate/core/" + name, configDir);
+        fr = TestUtils.runFeature("classpath:com/intuit/karate/core/" + name, configDir);
         if (fail) {
             assertTrue(fr.result.isFailed());
         } else {
