+++ karate-core/src/test/java/com/intuit/karate/core/HttpMockHandlerRunner.java
@@ -23,14 +23,12 @@
  */
 package com.intuit.karate.core;
 
+import static com.intuit.karate.TestUtils.*;
 import com.intuit.karate.http.ArmeriaHttpClient;
-import com.intuit.karate.match.Match;
-import com.intuit.karate.match.MatchResult;
 import com.intuit.karate.http.HttpRequestBuilder;
 import com.intuit.karate.http.HttpServer;
 import com.intuit.karate.http.Response;
 import org.junit.jupiter.api.AfterEach;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -63,11 +61,6 @@ class HttpMockHandlerRunner { // TODO investigate intermittent CI failure
         return mock;
     }
 
-    private void match(Object actual, Object expected) {
-        MatchResult mr = Match.that(actual).isEqualTo(expected);
-        assertTrue(mr.pass, mr.message);
-    }
-
     @AfterEach
     void afterEach() {
         server.stop();
@@ -75,7 +68,7 @@ class HttpMockHandlerRunner { // TODO investigate intermittent CI failure
 
     @Test
     void testProceed() {
-        FeatureBuilder fb = FeatureBuilder.background().scenario(
+        FeatureBuilder fb = background().scenario(
                 "pathMatches('/hello')",
                 "def response = 'world'");
         HttpServer downStream = new HttpServer(0, new MockHandler(fb.build()));
