+++ karate-core/src/test/java/com/intuit/karate/core/KarateHttpMockHandlerTest.java
@@ -1,11 +1,9 @@
 package com.intuit.karate.core;
 
-import com.intuit.karate.match.Match;
-import com.intuit.karate.match.MatchResult;
-import static com.intuit.karate.core.RuntimeUtils.runScenario;
+import static com.intuit.karate.TestUtils.*;
+import static com.intuit.karate.TestUtils.runScenario;
 import com.intuit.karate.http.HttpServer;
 import org.junit.jupiter.api.AfterEach;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -48,11 +46,6 @@ class KarateHttpMockHandlerTest {
 
     private void matchVar(String name, Object expected) {
         match(get(name), expected);
-    }
-    
-    private void match(Object actual, Object expected) {
-        MatchResult mr = Match.that(actual).isEqualTo(expected);
-        assertTrue(mr.pass, mr.message);
     }
 
     @AfterEach
