+++ karate-core/src/test/java/com/intuit/karate/core/ScenarioEngineTest.java
@@ -1,5 +1,6 @@
 package com.intuit.karate.core;
 
+import com.intuit.karate.TestUtils;
 import com.intuit.karate.StringUtils;
 import com.intuit.karate.Json;
 import com.intuit.karate.match.Match;
@@ -24,7 +25,7 @@ public class ScenarioEngineTest {
 
     @BeforeEach
     void beforeEach() {
-        engine = RuntimeUtils.engine();
+        engine = TestUtils.engine();
         engine.init();
     }
