+++ karate-core/src/test/java/com/intuit/karate/core/ScenarioRuntimeTest.java
@@ -3,7 +3,7 @@ package com.intuit.karate.core;
 import com.intuit.karate.FileUtils;
 import com.intuit.karate.match.Match;
 import com.intuit.karate.match.MatchResult;
-import static com.intuit.karate.core.RuntimeUtils.*;
+import static com.intuit.karate.TestUtils.*;
 import java.io.File;
 import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
